Update patch set 1
Patch Set 1: Code-Review+1

I'm happy to give it a +1 to resolve the issue with the scanning, but it does seem to me that these comments were meant more as a preemptive disclaimer than something that should have raised a red flag at all. I think the important person here is Eric, since he very carefully did the legwork to verify with legal that these images were usable.

Patch-set: 1
Label: Code-Review=+1